Routerings registraties wijzigen naar PHP8 attributes (#55)
